debug to allow different layermasks in each batch

+++ models/layer_mask_predictor.py
@@ -161,7 +161,7 @@ class LayerMaskPredictor(nn.Module):
             cut_offs = ((self.cut_offs - (lmp_input_mask == 0).sum(1).unsqueeze(1)) >= 0)
             bins = [(cut_off == 1).nonzero()[0] for cut_off in cut_offs]
             print("bins", bins)
-            return torch.stack([random.choice(self.combs_by_bin[b.item()]) for b in bins])
+            return torch.stack([random.choice(self.combs_by_bin[b.item()]) for b in bins]).float()
 
         if self.lmp_type == "ensemble":
             return self.layermasks  # .unsqueeze(0).expand(batch_size, -1)
